Create navigating notices, Change notice id parameter type (Int -> Long)

+++ data/src/main/java/daily/dayo/data/repository/NoticeRepositoryImpl.kt
@@ -19,7 +19,7 @@ class NoticeRepositoryImpl @Inject constructor(
             NoticePagingSource(noticeApiService, NOTICE_PAGE_SIZE)
         }.flow
 
-    override suspend fun requestDetailNotice(noticeId: Int): NetworkResponse<NoticeDetail> =
+    override suspend fun requestDetailNotice(noticeId: Long): NetworkResponse<NoticeDetail> =
         when (
             val response = noticeApiService.requestDetailNotice(noticeId)) {
             is NetworkResponse.Success -> NetworkResponse.Success(response.body?.toNoticeDetail())
